# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Monitoring API",
        "title": "Monitoring service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/api/login": {
            "post": {
                "description": "Login endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Login endpoint",
                "parameters": [
                    {
                        "description": "Dto для логина",
                        "name": "dto.LoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DataLoginResponse"
                        }
                    },
                    "default": {
                        "description": "The body of any response with an error",
                        "schema": {
                            "$ref": "#/definitions/error.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Logout endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Logout"
                ],
                "summary": "Logout endpoint",
                "responses": {
                    "200": {
                        "description": "message: Successfully logged out.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "The body of any response with an error",
                        "schema": {
                            "$ref": "#/definitions/error.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.DataLoginResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 3
                },
                "username": {
                    "type": "string",
                    "maxLength": 191
                }
            }
        },
        "error.HTTPError": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {},
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "statusCode": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}